Refactor namespaced interface
Encapsulate more of the logic for the namespace flags.
