+++ cmd/svcat/command/namespaced.go
@@ -16,6 +16,18 @@ limitations under the License.
 
 package command
 
+import "github.com/spf13/pflag"
+
+// HasNamespaceFlags represents a command that can be scoped to a namespace.
+type HasNamespaceFlags interface {
+	Command
+
+	// ApplyNamespaceFlags persists the namespace-related flags:
+	// * --namespace
+	// * --all-namespaces
+	ApplyNamespaceFlags(flags *pflag.FlagSet)
+}
+
 // Namespaced is the base command of all svcat commands that are namespace scoped.
 type NamespacedCommand struct {
 	*Context
@@ -27,12 +39,29 @@ func NewNamespacedCommand(cxt *Context) *NamespacedCommand {
 	return &NamespacedCommand{Context: cxt}
 }
 
-// GetContext retrieves the command's context.
-func (c *NamespacedCommand) GetContext() *Context {
-	return c.Context
 }
 
-// SetNamespace sets the effective namespace for the command.
-func (c *NamespacedCommand) SetNamespace(namespace string) {
-	c.Namespace = namespace
+// ApplyNamespaceFlags persists the namespace-related flags:
+// * --namespace
+// * --all-namespaces
+func (c *NamespacedCommand) ApplyNamespaceFlags(flags *pflag.FlagSet) {
+	c.Namespace = c.determineNamespace(flags)
+}
+
+// determineNamespace using the current context's namespace, and the user-requested namespace.
+func (c *NamespacedCommand) determineNamespace(flags *pflag.FlagSet) string {
+	currentNamespace := c.Context.App.CurrentNamespace
+
+	namespace, _ := flags.GetString("namespace")
+	allNamespaces, _ := flags.GetBool("all-namespaces")
+
+	if allNamespaces {
+		return ""
+	}
+
+	if namespace != "" {
+		return namespace
+	}
+
+	return currentNamespace
 }
